pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore APi"
authors = ["edumarques7 <marques.barbosa@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-rest-framework = "^0.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
